Add extract_lineage function
- Added extract lineage function, which creates a data structure similar to the comment tree, but in a way that's easier to search later on. Specifically, it creates a dictionary with integer keys and List values. The lists, as currently implemented, contain the IDs of the comments in the order up to and including a comment with a given ID, starting from a first-level comment child. In other words, given a comment tree that looks like this:
1 -> 2 -> 3, where the arrows indicate a reply, the dictionary D returned by extract_lineage would have the following at key 3:
D[3] = [1, 2, 3].
I think that this might be a better data structure than the original Tree data structure that was being constructed, because it will be much easier to update and/or search for items. This approach is likely more space-intensive that my original approach (with or without assuming relatively light space overhead for a Tree object), but the added speed benefit may be worth it.
- I want to compare the performance of the Tree approach and the extract_lineage approach time-wise so I can see what seems to work best, especially on larger comment trees.