Further implementation of extract_post_item_text
- I've added code that extract the direct <tr> elements of the fatitem <table> element's <tbody> child.
- I've also added code that determines whether the post in question is active or not based on the presence or absence of a comment form in the one of the <tr> elements extracted from the aforementioned <tbody> element.

The next step in the implementation is to use the combined knowledge of whether a form is active or not with the knowledge of the number of <tr> children to determine where to grab the text from. I learned that each tag has a .text property that is very handy to call in order to get this raw text, which can then likely be passed to the extract_comment_text method (which may need to be renamed to something more suitable, as it extracts more than just the text of comments).