Add MVP version of application
This commit contains code that enables a basic level of interaction with
the underlying data/object model created in the earlier part of the project.

A user can browse various posts on HN based on the ID of the post and the
rank of the post (if applicable, meaning the post is on one of HN's main
news pages). Additionally, users can save the last post they read, and
see the ranked posts on different pages of HN, and easily quit the
application.

Working on this part helped inspire lots of ideas for what I want the
final version of the application to look/feel like. In fact, I realized
that I may not even need to use the rich library in this app after all,
which is obviously a big deal. I'll have to see what happens from here,
but the main thing I want to do is start building an ncurses-based
version of the final app.