Add colored output for news page and article titles; new TODO for prettify_string
- The output of items that show up on the news page are now colored in a randomly
deterministic way based on item ID, so that way, a post with a given ID will have
always have the same color. This might help someone who remembers that they saw
a post on the front page that was blue, but doesn't recall the title. Even if not,
it looks cool. :)
- Added a TODO/clarifying comment for the part of prettify_string that processes
italics. I realized I was not handling italics split over multiple paragraphs in
a proper way, due to some erroneous assumptions about how that works.