Fix comment tree indentation bug
For some reason, the comment indentation in the app started erroneously treating
the first top-level comment on a given post as the parent of all lower comments
in the thread, throwing off the indentation. This small change should fix it by
ensuring that the lineage tree is cleared past the indentation difference of the
current comment being examined and the comment directly below it.